Refactor

+++ src/LocalGistClient.cpp
@@ -85,22 +85,22 @@ void LocalGistClient::closeStream() {
 void LocalGistClient::process(ofSoundBuffer &input, ofSoundBuffer &output) {
   gist.processAudio(input.getBuffer(), bufferSize, nChannels, sampleRate);
 
-  scalarValues[static_cast<int>(AnalysisScalar::rootMeanSquare)] = gist.getValue(GIST_ROOT_MEAN_SQUARE);
-  scalarValues[static_cast<int>(AnalysisScalar::peakEnergy)] = gist.getValue(GIST_PEAK_ENERGY);
-  scalarValues[static_cast<int>(AnalysisScalar::zeroCrossingRate)] = gist.getValue(GIST_ZERO_CROSSING_RATE);
-  scalarValues[static_cast<int>(AnalysisScalar::spectralCentroid)] = gist.getValue(GIST_SPECTRAL_CENTROID);
-  scalarValues[static_cast<int>(AnalysisScalar::spectralCrest)] = gist.getValue(GIST_SPECTRAL_CREST);
+  scalarValues[AnalysisScalar::rootMeanSquare] = gist.getValue(GIST_ROOT_MEAN_SQUARE);
+  scalarValues[AnalysisScalar::peakEnergy] = gist.getValue(GIST_PEAK_ENERGY);
+  scalarValues[AnalysisScalar::zeroCrossingRate] = gist.getValue(GIST_ZERO_CROSSING_RATE);
+  scalarValues[AnalysisScalar::spectralCentroid] = gist.getValue(GIST_SPECTRAL_CENTROID);
+  scalarValues[AnalysisScalar::spectralCrest] = gist.getValue(GIST_SPECTRAL_CREST);
 //  scalarValues[static_cast<int>(AnalysisScalar::spectralFlatness)] = gist.getValue(GIST_SPECTRAL_FLATNESS);
 //  scalarValues[static_cast<int>(AnalysisScalar::spectralRollof)] = gist.getValue(GIST_SPECTRAL_);
 //  scalarValues[static_cast<int>(AnalysisScalar::spectralKurtosis)] = gist.getValue(GIST_K);
 //  scalarValues[static_cast<int>(AnalysisScalar::energyDifference)] gist.getValue(GIST_);
-  scalarValues[static_cast<int>(AnalysisScalar::spectralDifference)] = gist.getValue(GIST_SPECTRAL_DIFFERENCE);
+  scalarValues[AnalysisScalar::spectralDifference] = gist.getValue(GIST_SPECTRAL_DIFFERENCE);
 //  scalarValues[static_cast<int>(AnalysisScalar::spectralDifferenceHWR)] gist.getValue(GIST_SPEC);
   scalarValues[static_cast<int>(AnalysisScalar::complexSpectralDifference)] = gist.getValue(GIST_SPECTRAL_DIFFERENCE_COMPLEX);
   // SPECTRAL_DIFFERENCE_HALFWAY
 //  scalarValues[static_cast<int>(AnalysisScalar::highFrequencyContent)] = gist.getValue(GIST_HIGH_FREQUENCY_CONTENT);
   float pitchEstimate = gist.getValue(GIST_PITCH);
-  if (pitchEstimate < 5000.0 && pitchEstimate > 10.0) scalarValues[static_cast<int>(AnalysisScalar::pitch)] = pitchEstimate;
+  if (pitchEstimate < 5000.0 && pitchEstimate > 10.0) scalarValues[AnalysisScalar::pitch] = pitchEstimate;
   
 //  auto mfccs = gist.getMelFrequencyCepstralCoefficients();
 //  if (mfccs.size() != mfcc.size()) {
